snowflake-databricks-mastery/requirements.txt
# Snowflake and Databricks Mastery - Python Dependencies

# ==================== Core Dependencies ====================
# Snowflake
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1

# Databricks and Spark
databricks-connect==14.2.0
pyspark==3.5.0
delta-spark==3.0.0

# Machine Learning
mlflow==2.9.2
scikit-learn==1.3.2
xgboost==2.0.3

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# ==================== Utilities ====================
# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
configparser==6.0.0

# Logging and Monitoring
loguru==0.7.2
colorlog==6.8.0

# Data Validation
great-expectations==0.18.8
pydantic==2.5.3

# ==================== Database & Storage ====================
# Additional database connectors
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Cloud storage
boto3==1.34.14  # AWS S3
azure-storage-blob==12.19.0  # Azure Blob Storage
google-cloud-storage==2.14.0  # GCP Cloud Storage

# ==================== Development Tools ====================
# Jupyter
jupyter==1.0.0
jupyterlab==4.0.10
ipykernel==6.27.1
notebook==7.0.6

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# ==================== Data Visualization ====================
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# ==================== API & Web ====================
requests==2.31.0
httpx==0.26.0
fastapi==0.108.0
uvicorn==0.25.0

# ==================== Streaming ====================
confluent-kafka==2.3.0
kafka-python==2.0.2

# ==================== Additional Libraries ====================
# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Data formats
openpyxl==3.1.2  # Excel
xmltodict==0.13.0  # XML
beautifulsoup4==4.12.2  # HTML/XML parsing

# Cryptography
cryptography==41.0.7

# Retry logic
tenacity==8.2.3

# Progress bars
tqdm==4.66.1

# ==================== Optional (Install as needed) ====================
# dbt (data build tool)
# dbt-core==1.7.4
# dbt-snowflake==1.7.1
# dbt-databricks==1.7.3

# Airflow (for orchestration)
# apache-airflow==2.8.0
# apache-airflow-providers-snowflake==5.1.0
# apache-airflow-providers-databricks==5.0.0

# Terraform Python client
# python-terraform==0.10.1
